Move VerifyGroupNumbers to cfg
